# openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce API
  description: |
    A comprehensive e-commerce REST API with authentication, product management, shopping cart, and order processing.

    ## Features
    - User authentication (register/login)
    - Product catalog with categories
    - Shopping cart management
    - Order processing and tracking
    - Admin dashboard and management

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@keystore.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.keystore.com/api/v1
    description: Production server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new customer account
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
                password:
                  type: string
                  minLength: 6
                  example: SecurePass123
            examples:
              customer:
                summary: Customer registration
                value:
                  name: Jane Smith
                  email: jane.smith@example.com
                  password: MyPassword123
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              examples:
                success:
                  value:
                    success: true
                    message: User registered successfully
                    data:
                      user:
                        id: 5
                        name: Jane Smith
                        email: jane.smith@example.com
                        role: customer
                        created_at: "2025-01-15T10:30:00.000Z"
                      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjUsInJvbGUiOiJjdXN0b21lciIsImp0aSI6ImFiYzEyMyIsImlhdCI6MTcwNTMxNzAwMCwiZXhwIjoxNzA1NDAzNDAwfQ.signature
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and receive JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@keystore.com
                password:
                  type: string
                  example: password123
            examples:
              customer:
                summary: Customer login
                value:
                  email: agus@example.com
                  password: password123
              admin:
                summary: Admin login
                value:
                  email: admin@keystore.com
                  password: password123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              examples:
                success:
                  value:
                    success: true
                    message: Login successful
                    data:
                      user:
                        id: 1
                        name: Admin User
                        email: admin@keystore.com
                        role: admin
                        created_at: "2025-01-01T00:00:00.000Z"
                      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsInJvbGUiOiJhZG1pbiIsImp0aSI6Inh5ejc4OSIsImlhdCI6MTcwNTMxNzAwMCwiZXhwIjoxNzA1NDAzNDAwfQ.signature
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      description: Retrieve list of all product categories
      operationId: getCategories
      responses:
        "200":
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
              examples:
                success:
                  value:
                    success: true
                    data:
                      - id: 1
                        name: Mechanical Keyboards
                      - id: 2
                        name: Keycaps
                      - id: 3
                        name: Switches
                      - id: 4
                        name: Accessories

    post:
      tags:
        - Categories
      summary: Create category
      description: Create a new product category (Admin only)
      operationId: createCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: Gaming Mice
            examples:
              category:
                value:
                  name: Desk Mats
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Category created successfully
                  data:
                    $ref: "#/components/schemas/Category"
              examples:
                success:
                  value:
                    success: true
                    message: Category created successfully
                    data:
                      id: 5
                      name: Desk Mats
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /categories/{id}:
    put:
      tags:
        - Categories
      summary: Update category
      description: Update category name (Admin only)
      operationId: updateCategory
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Category ID
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: Premium Mechanical Keyboards
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Category updated successfully
                  data:
                    $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags:
        - Categories
      summary: Delete category
      description: Delete a category (Admin only)
      operationId: deleteCategory
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Category ID
          schema:
            type: integer
          example: 5
      responses:
        "200":
          description: Category deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Category deleted successfully
                  data:
                    type: null
                    example: null
        "400":
          description: Cannot delete category with products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Referenced resource not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...